backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
pyotp==2.9.0
python-dotenv==1.0.0
cryptography==41.0.7

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Task Queue
celery==5.3.4
kombu==5.3.4

# Kafka
aiokafka==0.10.0

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0
faker==21.0.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security Scanning
bandit==1.7.6
# safety removed due to dependency conflicts
